+++ src/components/hooks/useCustomization.tsx
@@ -14,6 +14,7 @@ import SelectCard from '../Inspector/SelectCard';
 import Inspector from '../Inspector';
 import ThumbIndex from '../ThumbIndex';
 import Context from '../Context';
+import TextContext from '../Text/context';
 import {ConfigField, ConfigCard, useDragging} from '../Form/DragDrop';
 import {Button} from '../prime';
 import testid from '../lib/testid';
@@ -34,6 +35,7 @@ function getLayout(cards: Cards, layouts: Layouts, mode: LayoutMode, name = '') 
     const orientation = items?.orientation;
     const type = items?.type;
     const disableBack = items?.disableBack;
+    const hideBack = items?.hideBack;
     const disabled = items?.disabled;
     const enabled = items?.enabled;
     if (orientation) items = items.items;
@@ -53,6 +55,7 @@ function getLayout(cards: Cards, layouts: Layouts, mode: LayoutMode, name = '') 
         toolbar,
         layoutName,
         disableBack || false,
+        hideBack || false,
         disabled,
         enabled
     ];
@@ -88,6 +91,7 @@ export default function useCustomization({
 }) {
     const [inspected, onInspect] = React.useState(null);
     const {customization: customizationEnabled} = React.useContext(Context);
+    const {joiMessages, translate} = React.useContext(TextContext);
     const [design, toggleDesign] = useToggle(designDefault);
     const [mergedCustomization, setCustomization] = React.useState({schema: {}, card: {}, layout: {}, ...customizationDefault});
     const mergedSchema = React.useMemo(() => merge({}, schema, mergedCustomization.schema), [schema, mergedCustomization.schema]);
@@ -95,7 +99,7 @@ export default function useCustomization({
     const mergedLayouts = React.useMemo(() => merge({}, layouts, mergedCustomization.layout), [layouts, mergedCustomization.layout]);
     const [addField, setAddField] = React.useState(null);
     const [addCard, setAddCard] = React.useState(null);
-    const [items, layout, indexType, orientation, toolbar, currentLayoutName, disableBack, disabled, enabled] = React.useMemo(
+    const [items, layout, indexType, orientation, toolbar, currentLayoutName, disableBack, hideBack, disabled, enabled] = React.useMemo(
         () => getLayout(mergedCards, mergedLayouts, mode, layoutState),
         [mergedCards, mergedLayouts, mode, layoutState]
     );
@@ -386,28 +390,28 @@ export default function useCustomization({
         dropdownNames,
         layoutFields
     ] = React.useMemo(() => {
-        const {fields, validation, dropdownNames} = fieldNames(indexCards(items) || layoutItems || [], mergedCards, mergedSchema, editors);
+        const {fields, validation, dropdownNames} = fieldNames(indexCards(items) || layoutItems || [], mergedCards, mergedSchema, editors, translate);
         return [
             validation,
             dropdownNames,
             fields
         ];
-    }, [mergedCards, editors, items, layoutItems, mergedSchema]);
+    }, [mergedCards, editors, items, layoutItems, mergedSchema, translate]);
 
     const getLayoutValue = React.useCallback((mode, layoutState, value) => {
         const [items, layout] = getLayout(mergedCards, mergedLayouts, mode, layoutState);
         const layoutItems = items ? false : layout;
-        const {fields} = fieldNames(indexCards(items) || layoutItems || [], mergedCards, mergedSchema, editors);
+        const {fields} = fieldNames(indexCards(items) || layoutItems || [], mergedCards, mergedSchema, editors, translate);
         return getFieldsValue(fields, value);
-    }, [editors, mergedCards, mergedLayouts, mergedSchema]);
+    }, [editors, mergedCards, mergedLayouts, mergedSchema, translate]);
 
     const [resolver, isPropertyRequired] = React.useMemo(() => {
-        const [validationSchema, requiredProperties] = getValidation(mergedSchema);
+        const [validationSchema, requiredProperties] = getValidation(mergedSchema, translate);
         return [
-            joiResolver(validation || validationSchema, {stripUnknown: true, abortEarly: false}),
+            joiResolver(validation || validationSchema, {stripUnknown: true, abortEarly: false, messages: joiMessages}),
             propertyName => requiredProperties.includes(propertyName)
         ];
-    }, [validation, mergedSchema]);
+    }, [validation, mergedSchema, joiMessages, translate]);
     const formApi = useForm({resolver});
 
     const validate = React.useCallback(selectedList => {
@@ -427,6 +431,7 @@ export default function useCustomization({
             loading={loading}
             validate={validate}
             disableBack={disableBack}
+            hideBack={hideBack}
             methods={methods}
             formApi={formApi}
         />