+++ src/components/lib/parseDictionaryMap.ts
@@ -1,0 +1,5 @@
+export default content => content?.trim().split(/\r?\n/).reduce((prev, cur, index) => {
+    const [dictionaryKey, translatedValue] = cur.trim().split('=');
+    if (dictionaryKey && translatedValue) prev[dictionaryKey] = translatedValue;
+    return prev;
+}, {});